backend/requirements.txt
fastapi
uvicorn
google-genai
gTTS
